Rename setuid to st_mode, add files for F29, F30, and F31.
The files in this directory capture expected st_mode values for the
purposes of verying setuid and setgid files.  The subdirectory name was
setuid which is mostly correct.  Just call it st_mode.

Also add some data files in here for Fedora 29, 30, and 31.  These are
copies of the rawhide file.