Update fedora.yaml from generic.yaml, disable 'patches' inspection
From this report:
    https://github.com/rpminspect/rpminspect/issues/257

In Fedora the patches inspection may not make sense in the common
case.  Fedora changes frequently and that means patches will fluctuate
a lot.  The original intent of the patches inspection is to raise
warnings when large patches are added or removed or a lot of files are
modified.  This makes sense for releases where some level of API
compatibility is desired (e.g., RHEL or a stable Fedora release).

For at least the common case, I am disabling the patches inspection.
We can create a profile for 'fedora-stable' or something that enables
it but even that may be unnecessary as many updates to Fedora stable
releases are simply new versions of upstream projects.  Individual
packages can also enable the patches inspection in their own
rpminspect.yaml file if desired.