'make koji' skips branches that lack Koji build targets
The default operationg of 'make koji' now is to build for all
discovered branches that have Koji build targets.  Once a release is
EOL'ed, the targets in Koji go away but the code remains in dist-git.
Well, just skip those as we encounter them so that the default
operation of 'make koji' can be "build for everything".  Still allow
the $BRANCHES variable to override the set of branches to build.