Revert "javabytecode: set default version to 55"
This reverts commit 87c08e7e892346e34882ed0fc9a7541f884cc303.

The default JVM major version should always be 43 here because that
was the first JVM byte code major version number.  Having it set to 43
is a way to catch products that lack a JVM major version setting in
the config file.  For Fedora, we could set it to the first JVM byte
code major version ever shipped in Fedora.  But I don't know what that
is.

Signed-off-by: David Cantrell <dcantrell@redhat.com>