Include the list of icons that are provided by Adwaita icon theme
Using only the icons from https://specifications.freedesktop.org/icon-naming-spec/icon-naming-spec-latest.html
is not enough as there are packages such as geoclue2 that contains
desktop files that reference icons that are not listed in the
specification, but are provided by the default icon theme. Instead of
adding the icons to the list on case by case basis rather generate the
list of icons that are in the system. The list was generated from the
adwaita-icon-theme package which is a default icon theme in GNOME. Other
spins might want to extend the list if they have their own specific
icons.