Add /lib/firmware to the list of excluded paths for pathmigration
/lib/modules is there, /lib/firmware does the same thing.  I do not
understand why we are still packaging RPMs with /lib/* paths even
though we completed UsrMove and /lib is a symlink.  Seems like we
should have RPMs carrying modules in /usr/lib/modules and firmware in
/usr/lib/firmware as a result.

The end result is the same though.  The files are installed where we
expect them.  But since packages put things in /lib/modules and
/lib/firmware, it means the symlinks we have now can never go away.  I
find this odd considering how insistent we were on moving things from
/bin to /usr/bin, /sbin to /usr/sbin, /lib to /usr/lib, and /lib64 to
/usr/lib64.

¯\_(ツ)_/¯

Signed-off-by: David Cantrell <dcantrell@redhat.com>