Many policy and config updates:
1) In fedora.yaml, add product release regexps for Fedora releases up
to Fedora 50 and EL releases up to EL 15.  These need to be added each
time there's a new release with a dist tag.  They map to the files in
subdirectories in this package.

2) Add 'fc41' files for abi/, capabilities/, fileinfo/, icons/,
rebaseable/, and security/.  This was done by copying the fc40 rules
to fc41 and we can go from there.

3) Add the remedy/ subdirectory with vendor remedy strings.  This is
mostly provided as an example, but if you want to improve the remedy
strings specific to Fedora, this is where to do it.  The defaults in
rpminspect are meant to be more generic.  This remedy/fedora.toml file
contains copies of the default strings right now, but feel free to
modify them to make them more specific to Fedora and Fedora's
processes.

4) Add missing udevrules inspection block in fedora.yaml as well as
listing the inspection by default.

Signed-off-by: David Cantrell <dcantrell@redhat.com>