Add placeholder rule files for Fedoras 43 through 50
rpminspect needs rule files per version.  The idea is that anyone in
Fedora can send a PR to this project to start those rule files.  They
are not automatically created.  Occassionally this happens, but not
always.

I am adding placeholder rule files for Fedora 43 through 50 to
somewhat futureproof things for a while.

Fedora devs: Please remember, anyone can submit a PR to this project
to change rules or create new ones!

Resolves: #63